RayScatteringScene renamed and inherit from Scene
