+++ pytissueoptics/rayscattering/source.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from pytissueoptics.rayscattering.tissues import Tissue
+from pytissueoptics.rayscattering.tissues.rayScatteringScene import RayScatteringScene
 from pytissueoptics.scene import Vector, Material
 from pytissueoptics.rayscattering.photon import Photon
 from pytissueoptics.scene.intersection import SimpleIntersectionFinder
@@ -15,9 +15,9 @@ class Source:
 
         self._photons = photons
 
-    def propagate(self, tissue: Tissue, worldMaterial: Material = Material(), logger: Logger = None):
-        intersectionFinder = SimpleIntersectionFinder(tissue.getSolids())
-        tissue.setWorldMaterial(worldMaterial)
+    def propagate(self, scene: RayScatteringScene, worldMaterial: Material = Material(), logger: Logger = None):
+        intersectionFinder = SimpleIntersectionFinder(scene.getSolids())
+        scene.setWorldMaterial(worldMaterial)
         for photon in self._photons:
             photon.setContext(worldMaterial, intersectionFinder=intersectionFinder, logger=logger)
             photon.propagate()
